requirements.txt
# Content Verification Service - Requirements for Python 3.12
# Updated for compatibility with Python 3.12

# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Science & ML Core
numpy
scipy
pandas
scikit-learn

# Natural Language Processing
nltk
joblib

# Computer Vision
torch
torchvision
Pillow
opencv-python-headless

# Utilities
python-dotenv
pydantic
pydantic-settings

# Development
pytest
black
flake8

# Production
gunicorn